1625-cbkland-nft-bridge-proxy

+++ contracts/CBKLandBridgeProxyContract.sol
@@ -43,7 +43,7 @@ contract CBKLandBridgeProxyContract is Initializable, AccessControlUpgradeable, 
 
         uintVars = new uint256[](2);
         uintVars[UINT_NFT_VAR_META] = _packCBKLandData(tier, chunkId, x, y);
-        uintVars[UINT_NFT_VAR_ADDRESS] = uint256(reseller);
+        uintVars[UINT_NFT_VAR_ADDRESS] = uint256(uint160(reseller));
     }
 
     // for future use, bot will probe the returned value to know if the proxy contract has proper signature behavior
@@ -63,7 +63,7 @@ contract CBKLandBridgeProxyContract is Initializable, AccessControlUpgradeable, 
         require(_enabled, "not enabled");
 
         (uint256 tier, uint256 chunkId, uint256 x, uint256 y) = _unpackCBKLandData(uintVars[UINT_NFT_VAR_META]);
-        address reseller = address(uintVars[UINT_NFT_VAR_ADDRESS]);
+        address reseller = address(uint160(uintVars[UINT_NFT_VAR_ADDRESS]));
 
          tokenId = 
             _cbkLand.mintOrUpdate(tokenId, _nftStorageAddress, tier, chunkId, x, y, reseller);
@@ -76,13 +76,13 @@ contract CBKLandBridgeProxyContract is Initializable, AccessControlUpgradeable, 
     // x => max is 5000 => 16 bits more than enough
     // y => max is 5000 => 16 bits more than enough
     function _packCBKLandData(uint256 tier, uint256 chunkId, uint256 x, uint256 y) internal pure returns (uint256) {
-        return  uint256(tier | (uint256(chunkId) << 8) | (uint256(x) << 24) | (uint256(y) << 40));
+        return  tier | chunkId << 8 | x << 24 | y << 40;
     }
 
     function _unpackCBKLandData(uint256 metaData) internal pure returns (uint256 tier, uint256 chunkId, uint256 x, uint256 y) {
-        tier = uint8((metaData) & 0xFF);
-        chunkId = uint8((metaData >> 8) & 0xFFFF);
-        x = uint16(metaData  >> 24 & 0xFFFF);
-        y = uint32((metaData >> 40) & 0xFFFF);
+        tier = metaData & 0xFF;
+        chunkId = (metaData >> 8) & 0xFFFF;
+        x = (metaData  >> 24) & 0xFFFF;
+        y = (metaData >> 40) & 0xFFFF;
     }
 }
